requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
pydantic==2.9.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
openai>=1.40.0
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.2.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=24.1.0
pyyaml>=6.0
rdkit-pypi>=2022.9.5,<2023
rich>=13.0
python-dotenv>=1.0.1
